refactor breadcrumbs to its own component
