+++ frontend/src/components/navigation/Breadcrumbs.module.scss
@@ -1,19 +1,53 @@
 @import "styles/typography";
 @import "styles/colors";
+@import "styles/rules";
 @import "styles/spacings";
 
 .root {
   display: flex;
+  height: 50px;
+  padding: 0 $spacing-base - 2px;
+  border-radius: $border-radius-card;
   align-items: center;
+  background: $medium-violet;
+  margin: 0 $spacing-l;
 
-  & span {
-    margin: 0 $spacing-base * 0.5;
+  .backButtonWrapper {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: rgba(255, 255, 255, 0);
+    cursor: pointer;
+
+    .backButton {
+      width: 20px;
+      height: 16px;
+    }
+
+    transition: background 0.1s;
+
+    &:hover {
+      background: rgba(255, 255, 255, 0.12);
+    }
   }
 
-  * {
-    font-family: $font-family-inter;
-    font-size: $font-size-small - 1px;
-    color: $color-grey;
-    text-decoration: none;
+  .breadcrumbs {
+    margin-left: $spacing-base - 2px;
+    display: flex;
+    align-items: center;
+
+    & span {
+      margin: 0 $spacing-base * 0.5;
+    }
+
+    * {
+      font-family: $font-family-inter;
+      font-size: $font-size-small - 1px;
+      color: $color-grey;
+      text-decoration: none;
+    }
   }
 }
